+++ core/src/logic/Vegetable.java
@@ -9,10 +9,10 @@ import com.mygdx.game.Game;
  */
 public class Vegetable extends Element {
 
-    public static float default_radius = 5;
-    private float decayRate = 0;
+    public static double default_radius = 5;
+    private double decayRate = 0;
 
-    public Vegetable(float x, float y) {
+    public Vegetable(double x, double y) {
         super(x, y, default_radius);
     }
 
@@ -27,14 +27,14 @@ public class Vegetable extends Element {
     @Override
     public void render(ShapeRenderer s) {
         s.setColor(1, 1, 1, 1);
-        s.circle(getX(), getY(), getSize());
+        s.circle((float) getX(), (float) getY(), (float) getSize());
     }
 
-    public float getDecayRate() {
+    public double getDecayRate() {
         return decayRate;
     }
 
-    public void setDecayRate(float decayRate) {
+    public void setDecayRate(double decayRate) {
         this.decayRate = decayRate;
     }
 }
